tests moved to separate package
